prompt_api/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
pydantic
python-dotenv
celery[redis]
litellm
redis>=4.2.0 # Required for Sentinel support in celery